# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: TechEnglish Pro API
  version: 1.0.0
  description: |
    REST API contract cho hệ thống TechEnglish Pro (KLCN028).

    ## Quy ước
    - Base path: `/api/v1`.
    - ID sử dụng UUID.
    - Thời gian sử dụng ISO 8601 UTC.
    - Endpoint có biểu tượng khóa yêu cầu JWT Bearer token.
    - Danh sách dùng phân trang theo `page` và `limit`.
    - Client không được gửi điểm khi nộp bài; server là nguồn sự thật cho chấm điểm.

    ## Vai trò
    - `admin`: quản trị tài khoản, danh mục và toàn bộ dữ liệu.
    - `teacher`: quản lý nội dung, câu hỏi, bài thi và xem báo cáo được phép.
    - `learner`: học bài, làm bài thi và xem dữ liệu cá nhân.
  contact:
    name: TechEnglish Pro
  license:
    name: Proprietary - Academic Project
servers:
  - url: /api/v1
    description: Cùng host với API deployment; local mặc định là http://localhost:3080/api/v1.
...
paths:
...
  /domains:
    get:
      tags: [Domains]
      summary: Danh sách lĩnh vực CNTT
      operationId: listDomains
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/ActiveOnly'
      responses:
        '200':
          description: Danh sách lĩnh vực.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [Domains]
      summary: Tạo lĩnh vực CNTT
      operationId: createDomain
      requestBody:
        $ref: '#/components/requestBodies/DomainWriteBody'
      responses:
        '201':
          description: Lĩnh vực đã tạo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /levels:
    get:
      tags: [Levels]
      summary: Danh sách cấp độ học
      operationId: listLevels
      security: []
      responses:
        '200':
          description: Danh sách cấp độ theo thứ tự tăng dần.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [Levels]
      summary: Tạo cấp độ học
      operationId: createLevel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LevelWriteRequest'
      responses:
        '201':
          description: Cấp độ đã tạo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /career-goals:
    get:
      tags: [Career Goals]
      summary: Danh sách mục tiêu nghề nghiệp
      operationId: listCareerGoals
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/ActiveOnly'
      responses:
        '200':
          description: Danh sách mục tiêu nghề nghiệp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerGoalListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [Career Goals]
      summary: Tạo mục tiêu nghề nghiệp
      operationId: createCareerGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CareerGoalWriteRequest'
      responses:
        '201':
          description: Mục tiêu nghề nghiệp đã tạo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerGoalResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /certificates:
    get:
      tags: [Certificates]
      summary: Danh sách chứng chỉ
      operationId: listCertificates
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: domainId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/ActiveOnly'
      responses:
        '200':
          description: Danh sách chứng chỉ.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [Certificates]
      summary: Tạo chứng chỉ
      operationId: createCertificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateWriteRequest'
      responses:
        '201':
          description: Chứng chỉ đã tạo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /certificates/{certificateId}:
    parameters:
      - $ref: '#/components/parameters/CertificateId'
    get:
      tags: [Certificates]
      summary: Chi tiết chứng chỉ
      operationId: getCertificate
      security: []
      responses:
        '200':
          description: Chi tiết chứng chỉ.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Certificates]
      summary: Cập nhật chứng chỉ
      operationId: updateCertificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateWriteRequest'
      responses:
        '200':
          description: Chứng chỉ sau cập nhật.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Certificates]
      summary: Ngừng sử dụng chứng chỉ
      operationId: archiveCertificate
      responses:
        '204':
          description: Chứng chỉ đã được lưu trữ.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      description: Số trang, bắt đầu từ 1.
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Số phần tử mỗi trang.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Search:
      name: search
      in: query
      description: Tìm kiếm không phân biệt hoa thường theo các trường văn bản chính.
      schema:
        type: string
        minLength: 1
        maxLength: 100
...
    ActiveOnly:
      name: activeOnly
      in: query
      schema:
        type: boolean
        default: true
...
    CertificateId:
      name: certificateId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  requestBodies:
    DomainWriteBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DomainWriteRequest'
  responses:
...
    Forbidden:
      description: Người dùng không có quyền thực hiện thao tác.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Không tìm thấy tài nguyên.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Xung đột trạng thái hoặc dữ liệu đã tồn tại.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Dữ liệu đầu vào không hợp lệ.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    PaginationMeta:
      type: object
      required: [page, limit, totalItems, totalPages, hasNextPage, hasPreviousPage]
      properties:
        page: { type: integer, minimum: 1, example: 1 }
        limit: { type: integer, minimum: 1, example: 20 }
        totalItems: { type: integer, minimum: 0, example: 42 }
        totalPages: { type: integer, minimum: 0, example: 3 }
        hasNextPage: { type: boolean, example: true }
        hasPreviousPage: { type: boolean, example: false }
    ErrorDetail:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          description: Mã ổn định để client xử lý.
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          description: Thông báo an toàn cho người dùng hoặc developer.
          example: The requested resource was not found.
        requestId:
          type: string
          description: ID tương quan để tra log; không chứa dữ liệu nhạy cảm.
          example: req_01JABCDEF
    FieldViolation:
      type: object
      required: [field, message]
      properties:
        field: { type: string, example: email }
        code: { type: string, example: INVALID_FORMAT }
        message: { type: string, example: Email must be valid. }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ValidationErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          allOf:
            - $ref: '#/components/schemas/ErrorDetail'
            - type: object
              required: [details]
              properties:
                details:
                  type: array
                  items:
                    $ref: '#/components/schemas/FieldViolation'
...
    Domain:
      type: object
      required: [id, code, name, description, isActive, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, pattern: '^[A-Z0-9_]{2,30}$', example: CLOUD }
        name: { type: string, minLength: 2, maxLength: 100, example: Cloud Computing }
        description: { type: string, maxLength: 500 }
        icon: { type: string, nullable: true, maxLength: 100, example: cloud }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DomainWriteRequest:
      type: object
      additionalProperties: false
      required: [code, name, description]
      properties:
        code: { type: string, pattern: '^[A-Z0-9_]{2,30}$', example: DEVOPS }
        name: { type: string, minLength: 2, maxLength: 100, example: DevOps }
        description: { type: string, minLength: 1, maxLength: 500 }
        icon: { type: string, nullable: true, maxLength: 100 }
        isActive: { type: boolean, default: true }
    DomainResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: '#/components/schemas/Domain' }
    DomainListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, enum: [true] }
        data: { type: array, items: { $ref: '#/components/schemas/Domain' } }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
    Level:
      type: object
      required: [id, code, name, order, description, isActive]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, enum: [beginner, intermediate, advanced, professional] }
        name: { type: string, minLength: 2, maxLength: 50 }
        order: { type: integer, minimum: 1, maximum: 100 }
        description: { type: string, maxLength: 500 }
        isActive: { type: boolean }
    LevelWriteRequest:
      type: object
      additionalProperties: false
      required: [code, name, order, description]
      properties:
        code: { type: string, enum: [beginner, intermediate, advanced, professional] }
        name: { type: string, minLength: 2, maxLength: 50 }
        order: { type: integer, minimum: 1, maximum: 100 }
        description: { type: string, minLength: 1, maxLength: 500 }
        isActive: { type: boolean, default: true }
    LevelResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: '#/components/schemas/Level' }
    LevelListResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: { type: array, items: { $ref: '#/components/schemas/Level' } }
    CareerGoal:
      type: object
      required: [id, code, name, description, isActive]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, pattern: '^[A-Z0-9_]{2,40}$', example: BACKEND_ENGINEER }
        name: { type: string, minLength: 2, maxLength: 100, example: Backend Engineer }
        description: { type: string, maxLength: 500 }
        isActive: { type: boolean }
    CareerGoalWriteRequest:
      type: object
      additionalProperties: false
      required: [code, name, description]
      properties:
        code: { type: string, pattern: '^[A-Z0-9_]{2,40}$' }
        name: { type: string, minLength: 2, maxLength: 100 }
        description: { type: string, minLength: 1, maxLength: 500 }
        isActive: { type: boolean, default: true }
    CareerGoalResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: '#/components/schemas/CareerGoal' }
    CareerGoalListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, enum: [true] }
        data: { type: array, items: { $ref: '#/components/schemas/CareerGoal' } }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
    Certificate:
      type: object
      required: [id, code, name, provider, description, domainIds, isActive, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, pattern: '^[A-Z0-9_-]{2,40}$', example: AWS-SAA }
        name: { type: string, minLength: 2, maxLength: 150, example: AWS Certified Solutions Architect – Associate }
        provider: { type: string, minLength: 2, maxLength: 100, example: Amazon Web Services }
        description: { type: string, maxLength: 1000 }
        examUrl: { type: string, format: uri, nullable: true, maxLength: 2048 }
        domainIds: { type: array, uniqueItems: true, items: { type: string, format: uuid } }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CertificateWriteRequest:
      type: object
      additionalProperties: false
      required: [code, name, provider, description, domainIds]
      properties:
        code: { type: string, pattern: '^[A-Z0-9_-]{2,40}$' }
        name: { type: string, minLength: 2, maxLength: 150 }
        provider: { type: string, minLength: 2, maxLength: 100 }
        description: { type: string, minLength: 1, maxLength: 1000 }
        examUrl: { type: string, format: uri, nullable: true, maxLength: 2048 }
        domainIds: { type: array, minItems: 1, uniqueItems: true, items: { type: string, format: uuid } }
        isActive: { type: boolean, default: true }
    CertificateResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: '#/components/schemas/Certificate' }
    CertificateListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, enum: [true] }
        data: { type: array, items: { $ref: '#/components/schemas/Certificate' } }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
...
security:
  - bearerAuth: []